requirements.txt
garth>=0.5.0,<0.6.0
loguru
PyYAML
